@@ -12,9 +12,9 @@
     "bootstrap-icons": "^1.11.3",
     "lodash": "^4.17.14",
     "popper.js": "^1.16.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-router-dom": "^4.3.1",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-router-dom": "^6.30.1",
     "react-scripts": "^5.0.1"
   },
   "scripts": {
